fix(ros-z-dds): resolve clippy warnings (collapsible_if, new_ret_no_self, field_reassign_with_default, unnecessary casts)

+++ crates/ros-z-dds/src/cyclors/reader.rs
@@ -33,9 +33,7 @@ impl RawSample {
     }
 
     fn as_slice(&self) -> &[u8] {
-        unsafe {
-            slice::from_raw_parts(self.data.iov_base as *const u8, self.data.iov_len as usize)
-        }
+        unsafe { slice::from_raw_parts(self.data.iov_base as *const u8, self.data.iov_len) }
     }
 }
